[SPRINT_A]package refactoring
