+++ MoneyCircle/app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:label="@string/activity_register_and_SignIn" >
         </activity>
         <activity
-            android:name=".SetSplitAmount"
+            android:name=".split.SetSplitAmount"
             android:label="@string/title_activity_set_split_amount" >
         </activity>
         <activity
